requirements.txt
django==3.2
django-filter==2.4.0
django-environ==0.4.5
drf-yasg==1.20.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary
django-cors-headers==4.0.0
python-dotenv~=1.0.0
firebase-admin
requests